requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
pydantic-settings>=2.2
pyyaml>=6.0
scikit-learn>=1.3
httpx>=0.27
